Revert RRef.to_here()/local_value() return type
The return types of RRef.to_here()/local_value() were recently
changed to Future, which triggers flakiness as the RRef could be
deleted before the future.wait() finishes. While we are still
discussing how we'd like to solve it, this commit reverts the
return type to stop bleeding in tests.

ghstack-source-id: 7ae2d296f996cf4627303aadc230312dbd188e22
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29396